Finishing comments on the font, Cleaning up front code, fixing and adding some functionalities

+++ client/src/components/AmountPopUp/AmountPopUp.tsx
@@ -11,20 +11,21 @@ export interface CategoryPopUpProps {
     back: () => void
 }
 
+// Formats the date to be compatible with the date input type
 function formatDate(date: Date) {
     return `${date.getUTCFullYear()}-${( '0' + (date.getUTCMonth()+1) ).slice(-2)}-${( '0' + date.getUTCDate() ).slice(-2)}`;
 }
 
 export default function AmountPopUp({
-    currentCategory,
-    amount,
-    close,
-    refresh,
-    back
+    currentCategory, // The current category (used to determine if the created amount has a parent or not)
+    amount, // Optional, the amount that needs editing
+    close, // A function called when the pop up needs to close itself
+    refresh, // A function called when the pop up changed some things on the server, and the page probably needs a refresh
+    back // A function called when the pop up needs to go back one category in the category tree
 }: CategoryPopUpProps) {
 
     // The data of the current edited amount (not used if creating an amount)
-    const [ amountData, setAmountData ] = useState<Amount>({ amount: 0, dateTime: 0, description: '', gain: false, planned: 'no', uuid: '' });
+    const [ amountData, setAmountData ] = useState<Amount>({ amount: 0, dateTime: -1, description: '', gain: false, planned: false, uuid: '' });
 
     // If the amount prop is set, we fetch the amount that needs editing and we set the amountData state to the response from the server
     useEffect(() => {
@@ -34,6 +35,7 @@ export default function AmountPopUp({
             .catch((err) => console.error(err));
     }, []);
 
+    // A ref to the form, used to get back values from it
     const amountFormRef = createRef<HTMLFormElement>();
 
     // When the component is in editing mode this function is called when the form is submited
@@ -43,7 +45,6 @@ export default function AmountPopUp({
         const formData = {
             ...amountData,
             uuid: amount,
-            gain: amountData.gain ? 'true' : 'false'
         };
         
         // It fetches the api to edit this amount, and when the editing is done, it refreshes the page, and closes the pop up
@@ -61,20 +62,23 @@ export default function AmountPopUp({
             .catch((err) => console.error(err));
     };
 
+    // When the component is in creating mode this function is called when the form is submited
     const createAmount = (e: MouseEvent) => {
         e.preventDefault();
         if(!amountFormRef.current) return;
+        // It gets the form data, extracts the needed values from it and adds the parent category if one is present
         const form = new FormData(amountFormRef.current);
         
         const formData = {
             amount: Number(form.get('amount')),
-            gain: form.get('gain') === 'on' ? 'true' : 'false',
-            planned: form.get('planned') === 'on' ? 'monthly' : 'no',
+            gain: form.get('gain') === 'on',
+            planned: form.get('planned') === 'on',
             dateTime: Date.parse((form.get('date') || '').toString()) || -1,
             description: form.get('description'),
             category: currentCategory?.uuid || null
         };
         
+        // It fetches the api to create this amount, and when the creation is done, it refreshes the page, and closes the pop up
         fetch('/api/amount', {
             method: 'POST',
             headers: {
@@ -89,10 +93,15 @@ export default function AmountPopUp({
             .catch((err) => console.error(err));
     };
 
+    // Function used to delete the currently edited amount
     const deleteCurrentAmount = () => {
+        // Asks for confirmation
+        if(!confirm(`Are you sure you want to delete this amount\nThis action is irreversible`)) return;
+        // Deletes the current amount and refreshes the page
         fetch(`/api/amount/${amount}/`, {
             method: 'DELETE'
         }).then(() => {
+            alert('Amount deleted');
             refresh();
             close();
         }).catch((err) => console.error(err));
@@ -100,13 +109,13 @@ export default function AmountPopUp({
 
     return (
         <div id='popup-back' className={ styles['creation-pop-up'] } onClick={ (e) => {
+            // If the background of the pop up is clicked, close the pop up, else do nothing
             if((e.target as HTMLElement).id !== 'popup-back') return;
             close();
         } }>
             <form className={ styles['creation-form'] } ref={ amountFormRef }>
                 <p>{ amount ? 'Edit' : 'Create new' } Amount</p>
                 <div>
-                    { JSON.stringify(amountData) }
                     <div>
                         <label htmlFor="amount">Amount</label>
                         <input type="number" name="amount" id="amount" value={ amountData.amount } onChange={ (e) => setAmountData((data) => { return { ...data, amount: Number(e.target.value) }; }) } />
@@ -117,9 +126,9 @@ export default function AmountPopUp({
                     </div>
                     <div>
                         <label htmlFor="planned">Is it planned</label>
-                        <input type="checkbox" name="planned" id="planned" checked={ amountData.planned === 'monthly' } onChange={ (e) => setAmountData((data) => { return { ...data, planned: e.target.checked ? 'monthly' : 'no' }; }) } />
+                        <input type="checkbox" name="planned" id="planned" checked={ amountData.planned } onChange={ (e) => setAmountData((data) => { return { ...data, planned: e.target.checked }; }) } />
                     </div>
-                    <div>
+                    <div> { /* The date of the new amount or the new date of the edited amount, date is formated into a DD-MM-YYYY format to be compatible with the date input type */ }
                         <label htmlFor="date">Date</label>
                         <input type="date" name="date" id="date" value={ formatDate(new Date(amountData.dateTime)) } onChange={ (e) => setAmountData((data) => { return { ...data, dateTime: Date.parse(e.target.value) }; }) } />
                     </div>
